Added support for AWS Profiles
In some cases, the default profile is not able to be used with temporary credentials issued through a login provider such as Okta or OneLogin. This update will check for available profiles and prompt the user to enter a profile name if more than one profile is available. If there is only one, the script will carry on as usual.