update resource-count-oci.sh - get_compartments
I found an issue using this script. In the 78 line the script will parse the compartment-id, and it appears that it creates and issue listing just the root compartment, considering each child compartment shares the same parent "compartment-id", we could see the script generating a list of "0 resources". 

The proposed change is to parse the child compartment id ("id" instead "compartment-id"), avoinding a loop just to the root compartment, because the script is looping against the root compartment ("compartment-id" of a child compartment) instead the child compartments ("id" key).